coding/requirements.txt
Flask==3.0.3
tinydb==4.8.2
python-dateutil==2.9.0.post0

